Revert any changes to existing test classes. Add test class that checks other test classes for Linux vs OSX test count. Move array shuffle extensions inside LinuxMain.swift.
